+++ mission-racing/src/racingSpring/src/main/java/com/example/racingspring/domain/vo/MoveCount.java
@@ -1,4 +1,4 @@
-package com.example.racingspring.vo;
+package com.example.racingspring.domain.vo;
 
 public class MoveCount {
