+++ mission-racing/src/racingSpring/src/main/java/com/example/racingspring/repository/CarRepository.java
@@ -1,0 +1,4 @@
+package com.example.racingspring.repository;
+
+public class CarRepository {
+}
